requirements.txt
python-telegram-bot==20.7
httpx==0.25.2
fastapi==0.109.2
uvicorn[standard]==0.27.1
aiosqlite==0.20.0
apscheduler==3.10.4
jinja2==3.1.3
python-dotenv==1.0.0
yfinance>=0.2.40
pyyaml>=6.0
